+++ MyGame/Animations.h
@@ -22,7 +22,7 @@ typedef CAniFrame* LPANIMATION_FRAME;
 
 class CAnimation
 {
-	DWORD lastFrameTime;
+	ULONGLONG lastFrameTime;
 	int currentFrame;
 	int defaultTime;
 	vector<LPANIMATION_FRAME> frames;
